refactor: Extract HeaderInfoCard from PlannedTransactionsPage
