+++ CHANGELOG.md
@@ -38,3 +38,4 @@ refactoring was necessary:
   * Move `lang` and `dictRequire` to utils/lang
 * Ability to run the push notifications debug server from the CLI
 * Add some documentation to develop for android
+* Extract HeaderInfoCard from PlannedTransactionsPage
